# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: RBAC Platform API
  version: 0.1.0
servers:
  - url: http://localhost:8080
paths:
  /api/health:
    get:
      summary: Health check
      responses:
        '200':
          description: OK
  /api/auth/register:
    post:
      summary: Register user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '200':
          description: Register success
  /api/auth/login:
    post:
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
  /api/auth/refresh:
    post:
      summary: Refresh token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
      responses:
        '200':
          description: Refresh success
  /api/auth/logout:
    post:
      summary: Logout
      responses:
        '200':
          description: Logout success
  /api/users:
    get:
      summary: List users
      parameters:
        - in: query
          name: page
          schema:
            type: integer
        - in: query
          name: size
          schema:
            type: integer
      responses:
        '200':
          description: User list
    post:
      summary: Create user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        '200':
          description: User created
  /api/users/{id}:
    get:
      summary: Get user detail
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: User detail
    put:
      summary: Update user
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
      responses:
        '200':
          description: User updated
    delete:
      summary: Delete user
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: User deleted
  /api/users/{id}/status:
    patch:
      summary: Update user status
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserStatusRequest'
      responses:
        '200':
          description: User status updated
  /api/users/{id}/roles:
    put:
      summary: Bind roles to user
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BindRoleRequest'
      responses:
        '200':
          description: Roles bound
  /api/roles:
    get:
      summary: List roles
      responses:
        '200':
          description: Role list
    post:
      summary: Create role
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RoleCreateRequest'
      responses:
        '200':
          description: Role created
  /api/roles/{id}:
    put:
      summary: Update role
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RoleUpdateRequest'
      responses:
        '200':
          description: Role updated
    delete:
      summary: Delete role
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Role deleted
  /api/roles/{id}/permissions:
    put:
      summary: Bind permissions to role
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BindPermissionRequest'
      responses:
        '200':
          description: Permissions bound
  /api/permissions:
    get:
      summary: List permissions
      responses:
        '200':
          description: Permission list
    post:
      summary: Create permission
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PermissionCreateRequest'
      responses:
        '200':
          description: Permission created
  /api/permissions/{id}:
    put:
      summary: Update permission
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PermissionUpdateRequest'
      responses:
        '200':
          description: Permission updated
    delete:
      summary: Delete permission
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Permission deleted
  /api/logins:
    get:
      summary: List login logs
      parameters:
        - in: query
          name: page
          schema:
            type: integer
        - in: query
          name: size
          schema:
            type: integer
        - in: query
          name: username
          schema:
            type: string
      responses:
        '200':
          description: Login logs
  /api/operations:
    get:
      summary: List operation logs
      parameters:
        - in: query
          name: page
          schema:
            type: integer
        - in: query
          name: size
          schema:
            type: integer
        - in: query
          name: module
          schema:
            type: string
      responses:
        '200':
          description: Operation logs

components:
  schemas:
    RegisterRequest:
      type: object
      required: [username, email, password]
      properties:
        username:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
          format: password
    LoginRequest:
      type: object
      required: [account, password]
      properties:
        account:
          type: string
        password:
          type: string
          format: password
    LoginResponse:
      type: object
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        expiresIn:
          type: integer
    RefreshTokenRequest:
      type: object
      required: [refreshToken]
      properties:
        refreshToken:
          type: string
    CreateUserRequest:
      type: object
      required: [username, email, password]
      properties:
        username:
          type: string
        email:
          type: string
        password:
          type: string
        nickname:
          type: string
        phone:
          type: string
    UpdateUserRequest:
      type: object
      required: [email]
      properties:
        email:
          type: string
        nickname:
          type: string
        phone:
          type: string
    UpdateUserStatusRequest:
      type: object
      required: [status]
      properties:
        status:
          type: integer
    RoleCreateRequest:
      type: object
      required: [code, name]
      properties:
        code:
          type: string
        name:
          type: string
        description:
          type: string
    RoleUpdateRequest:
      type: object
      required: [code, name]
      properties:
        code:
          type: string
        name:
          type: string
        description:
          type: string
    BindPermissionRequest:
      type: object
      required: [permissionIds]
      properties:
        permissionIds:
          type: array
          items:
            type: integer
            format: int64
    BindRoleRequest:
      type: object
      required: [roleIds]
      properties:
        roleIds:
          type: array
          items:
            type: integer
            format: int64
    PermissionCreateRequest:
      type: object
      required: [code, name]
      properties:
        code:
          type: string
        name:
          type: string
        type:
          type: string
        resource:
          type: string
        method:
          type: string
        description:
          type: string
    PermissionUpdateRequest:
      type: object
      required: [code, name]
      properties:
        code:
          type: string
        name:
          type: string
        type:
          type: string
        resource:
          type: string
        method:
          type: string
        description:
          type: string